% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-translation-modules.R
\name{get_translation_modules}
\alias{get_translation_modules}
\title{Gets translation modules}
\usage{
get_translation_modules(package, language = Sys.getenv("LANGUAGE", "en"))
}
\arguments{
\item{package}{character string with the name of the package}

\item{language}{character string with the language}
}
\description{
Lists the package names of modules that translates a particular package
to a particular language
}
\keyword{internal}
